[FIX] knowledge: disable drag and drop on article tree for mobiles
Currently, the NestedSortable library may capture the swipe gesture made
over the article tree and may prevent the user from scrolling the page.
This behavior can be really annoying when the tree takes the whole screen.

To avoid that, we will disable the NestedSortable library when the user
uses a mobile device. That way, the swipe gesture will not be captured
and the user can scroll the page easily. The mobile users can use the
"Move To" modal to move an article.

task-2867059

Part-of: odoo/enterprise#37065