[IMP] l10n_co_edi: support for self-withholding taxes
Currently, all the taxes included in the invoices are being sent in the XML for
the electronic invoices.

There are a couple of common processes in Colombia that require the user to
include taxes in the invoices only for accounting purposes but do not need to be
reported in the XML. Like, Self-Withholding and StateWitholdings (Rete ICA).

So in this commit, In the Taxes value of the `Tipo de Valor` code has 'ZZ' and '07'
skip them on the XML creation for EDI.

TaskID - 2976963

closes odoo/enterprise#32803

X-original-commit: f23a6d55c5c79f9ede6ac68a701ce947cac3c952
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Bhavin Patel (bat) <bat@odoo.com>