[IMP] l10n_co_edi: support for self-withholding taxes
Currently, all the taxes included in the invoices are being sent in the XML for
the electronic invoices.

There are a couple of common processes in Colombia that require the user to
include taxes in the invoices only for accounting purposes but do not need to be
reported in the XML. Like, Self-Withholding and StateWitholdings (Rete ICA).

So in this commit, In the Taxes value of the `Tipo de Valor` code has 'ZZ' and '07'
skip them on the XML creation for EDI.

TaskID - 2976963

closes odoo/enterprise#32760

X-original-commit: e00821f398fb35a8b2b057c85a6cd54eec25a044
Signed-off-by: Josse Colpaert <jco@odoo.com>