[FIX] stock_barcode: convert dummy_id into number
When got from the fetched data, the `dummy_id` is a string.
This variable is used to match and link new lines without `id` with data
fetched from the server when these new lines are saved and in this case
the `dummy_id` is saved as a number in line's `virtual_id`.
The issue is we still can compare the select line's `virtual_id` with
its `dummy_id` when we want to get the selected line (`BarcodeModel`'s
`selectedLine`), and comparing a string and a number with a strict equal
can't work.

How to reproduce:
- Open a delivery in the Barcode App;
- Scan a product (add the new line if needed);
- Open the form view of the scanned product;
- Quit the form view by click the discard button or the back button;
- Scan again the same product -> Traceback.

It's caused because the application is supposed to have a selected line
(the scanned/edited one), so the model's `selectedLineVirtualId` has the
`dummy_id`/`virtual_id` of the selected line but while trying to find it
it compares the `selectedLineVirtualId` with the line's `dummy_id` and
as one is a number and another is a string, the line isn't returned.

X-original-commit: 59dea64f85435b77e7435f994463d6d6e7f3d0f4
Part-of: odoo/enterprise#34154