[FIX] knowledge: keep focus in embedded views
oe-blackbox (name may change in the future) is used by OdooEditor to prevent it
from interfering with the contents of the div which has that class.

In this case, it will allow to keep the focus in an input which has a parent
contenteditbale=false (i.e. searchbar in embedded views).

task-3004117

X-original-commit: 5a9cbaa9cbdaec1ab76e1b9e3a3b6e5eb6639d7a
Part-of: odoo/enterprise#32543