[FIX] website_helpdesk_form: wrong partner for same email user
Steps to reproduce:
- Create a partner "user 1" and then a user "user 2" with the same
email address "test@test.com"
- When connected as user 2 create a ticket

Current behavior:
The ticket is assigned to "user 1"

Expected behavior:
The ticket is assigned to "user 2"

Explanation:
To get the partner the code used to only take the first partner in the
list of partners with the mail written in the form. Now it will prioritise
the partner corresponding to the connected user if it is its email address.

opw-3007846

closes odoo/enterprise#33553

X-original-commit: 25db2bfe337e9f0f692639925e2d59e9834e574c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>