[IMP] knowledge: remove selection listener from the knowledge plugin
The editor no longer displays the text edition toolbar when the user
selects text located in an element having the class `oe-blackbox`.
As a result, we no longer need to handle text selection ourself. We
will therefore remove the dead selection listener from the Knowledge
plugin.

task-3004117

Part-of: odoo/enterprise#32327