[FIX] knowledge: fix main flow tour
Purpose:
--------
In the main flow tour, the "invite" button of the invite modal was clicked
before that the user to invite had been created, causing the tour to fail.

An extra_trigger has been added to make sure that the user has been created
before clicking on the invite button.

Task-3004122

Part-of: odoo/enterprise#32263