[FIX] knowledge: open newly created article when clicking on "create"
When the user opens the embedded views of the article items and clicks
on the "create" button, the user was redirected to the Knowledge editor
but the editor did not opened the newly created article. This commit will
ensure that the editor will open the newly created record after the user
clicks on the "create" button.

task-3004117

Part-of: odoo/enterprise#32327