[FIX] l10n_de_pos_cert: properly call super
When the german localisation is installed and we are not in a german
company using fiscal module. The override of the function
'_finalizeValidation' in Payment Screen does not correctly call the
parent method.

So we now check that we are in the german fiscal module conditions to
call what is specific to this localisation.

closes odoo/enterprise#32512

Signed-off-by: Masereel Pierre <pim@odoo.com>