[FIX] website_sale_dashboard: restore community ecommerce dashboard
Since [1] the community ecommerce dashboard was not rendered anymore.
This is because a `t-name` was combined with a `t-extend`.
This is unsure why but since `t-extend` is deprecated anyway, we replace
it by `t-inherit`.

This part is fixed in the community related PR, which then introduced
an issue with the template adapted here.
Note that for some reason, if the inherit mode is extension, a template
has to inherit the base one of the inherit hierarchy and not the one he
wants to inherit..

```xml
<!-- module A -->
<template t-name="A.template"/>
<!-- module B -->
<template t-name="B.template" t-inherit="A.template" t-inherit-mode="extension"/>
<!-- module C -->
<template t-name="C.template" t-inherit="A.template" t-inherit-mode="extension"/>
```

[1]: https://github.com/odoo/odoo/commit/dcf79f01b33a4f3ddb7e77480f0a7a8be1cd3794

task-2993773

closes odoo/enterprise#32464

Related: odoo/odoo#102286
Signed-off-by: Romain Derie (rde) <rde@odoo.com>