[FIX] account_reports: link report records to correct view
With the new reporting engine, the report lines, when clicked on,
would redirect to its basic form. As we already have a method to display
specific views for specific types of account moves and account move
lines, we now use them to get to the correct view.

This is done for the report lines in:
    - The journal report: now links to the correct form or widget for
    payments, bank statements, journal entries
    - All other reports: now links to the bank reconciliation widget
    when choosing the "View bank statement" caret option.

closes odoo/enterprise#32588

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>