[FIX] delivery_easypost: send state code instead of its name
The Easypost service does not work with some countries

To reproduce the issue:
1. Create a shipping method SM with Easypost-Fedex
2. Create a customer C:
    - City: Dispur
    - State: Assam
    - Code: 781005
    - Country: India
3. Create a SO:
    - Customer: C
    - Order Lines: add a product
4. Add Shipping:
    - Method: SM
    - Get the rate

Error: a User Error is displayed "It seems Easypost do not provide
shipments for this order [...] FedEx returned error: The length of the
destination state or province exceeds the limit of 2 characters."

The current code sends the state name instead of its code.

Easypost support contacted, answer "[...] In short, yes. It is
recommended to always provide a state/province 2-character abbreviation
rather than a full state/province name in order to avoid these types of
errors, regardless of country or carrier."

Locally tested with a client from:
- Assam (AS), in India
- Victoria (VIC), in Australia
- Pirkanmaa (FI-11), in Finland

OPW-3005900

closes odoo/enterprise#32942

X-original-commit: f58676a2a98f9c7597abda2a889d65dbe3cc57d7
Signed-off-by: Tiffany Chang <tic@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>