[IMP] spreadsheet_edition: new icon for global filters
SInce we introduced the data filters feature, we use the same icon
(fa-filter) for both data filters and global datasources filters.
For better distinction of the feature and keep a similar look to the
data filters of Gsheet, we change the icon of the global datasource
filters from `fa-filter` to `fa-sliders`.

X-original-commit: 7e16fd63d2620c950a15e5ff03168ad82dbb5f3b
Part-of: odoo/enterprise#32656