[IMP] voip: merge path and filename fields of SoundEffect model
Context:
We want to get rid of default values on identifying fields, and
SoundEffect/path is currently the only field to present this pattern.

Merging path and filename fields together makes the default value no
longer meaningful, as there is no full path that is shared by multiple
records and which would therefore consist of a relevant default value.

The point of removing default values from identifying fields is to
simplify the code of findFromIdentifyingData, which is called by every
single insert, so that we no longer have the overhead of looking for
default values.

This commit adapts the Enterprise code, and especially the VoIP code, to
the changes made to the SoundEffect model.

closes odoo/enterprise#32634

Related: odoo/odoo#102912
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>