[FIX] planning: selection badge layout in planning.slot form
Since the usage of the `CSS Grid` some views are broken.
The new form view system is less tolerant to missing colspan.
This commit, simplifies the `ARCH` to avoid over-customisation and uses
the framework system, fixing the problem at the same time.

Note: we remove the `XPath` as it's tha same field.

Steps to reproduce:
* Open Planning app
* Select the Kanban view
* Select a slot (e.g. Jeffrey Kelly - Community Manager) => Bug

closes odoo/enterprise#32718

X-original-commit: 3e930cdd38e1a2c82316920c957cbc59bca45cb7
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>