[IMP] *_extract: manual cron execution on refresh button click
If the worker crons are busy at the time of the trigger of the
parse cron, it won't be processed right away.
This makes the OCR feel slow as it can take multiple minutes
before a worker cron is available again to process the cron and
upload the documents to the server.

In this new version, the cron will be manually executed upon
clicking on the "Update" button when the document hasn't been
uploaded yet.

closes odoo/enterprise#32683

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>