[FIX] documents: fix invalid source for some files
The source for the attachment viewer could sometimes not work as you
might not always have direct access to an attachment but have to go
through the document instead.
This commit ensures we use the document instead of the attachment to get
the data to have a proper right access check.

TaskId-3015878

Part-of: odoo/enterprise#32713