[FIX] account_reports: fix regrouping of Accounts Coverage Report
Before this commit, if multiple account codes had the same error coming from different lines, they would all be regrouped into a single line in the Accounts Coverage Report with the same error. This is not ideal because in the case of a duplicate error, a unique error would regroup all the lines instead of only the real duplicates.

This commit fixes this by adding a condition on the regrouping to also check the lines.

closes odoo/enterprise#32745

X-original-commit: b62b981a733c4cfa9ebec5258842b34a69324cfc
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>