[FIX] account_sepa_direct_debit: don't display partner name in address
Currently, we are using computed field partner.contact_address to fill
in AddrLine element in SDD XML files. However, the company name is
included by default, which is not necessary since the company name is
already present in Nm Element just above. We should display the address
without the company name instead.

Before this commit:
        <Dbtr>
          <Nm>Azure Interior</Nm>
          <PstlAdr>
            <Ctry>US</Ctry>
            <AdrLine>Azure Interior 4557 De Silva St  Fremont CA 94538 United States</AdrLine>
          </PstlAdr>
        </Dbtr>

After this commit:
        <Dbtr>
          <Nm>Azure Interior</Nm>
          <PstlAdr>
            <Ctry>US</Ctry>
            <AdrLine>4557 De Silva St  Fremont CA 94538 United States</AdrLine>
          </PstlAdr>
        </Dbtr>

closes odoo/enterprise#33085

X-original-commit: 30b03543b6e492c1d492a1e7668ea1da0d35e7fe
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>