[REF] social_push_notifications: remove support of firebase_admin lib
This commit removes the support of the firebase_admin python library.

Indeed, it was not adding much value compared to the much simpler
implementation using the REST API.

The usage of this REST API still requires the "google_auth" python library to
be installed, but this one is much more common and available in debian
repositories by default.

Task-3000338

closes odoo/enterprise#32777

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>