[FIX] knowledge: make removeContextUserInfo a web asset
removeContextUserInfo was a Spreadsheet asset, which is not a dependency of
Knowledge.

In order to make this feature available for all modules, it is moved in web as
an object util, that needs to be used alongside the user service in order to
dynamically remove user context information (no hardcoding).

Task-3017349

closes odoo/enterprise#33322

X-original-commit: f8c00499a77b43b96f8577380948817b6a70e905
Related: odoo/odoo#104304
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>