[FIX] web_enterprise: list-style HomeMenu of iOS mobile app
Due to Apple AppStore's review, the HomeMenu in the mobile app on iPhone
should look like a list instead of the regular grid of icons.

This commit forces each row to take the whole width instead of being
cropped to the left side, hidding the App's name (but not the icon).

Steps to reproduce:
- Open your Odoo database on the iOS mobile app on iPhone
- Login
=> HomeMenu should look like a list but each row is cropped to the left

closes odoo/enterprise#32794

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>