[FIX] industry_fsm_stock: task assignee becoming salesman
Steps to reproduce:
- install sale_management and industry_fsm_stock;
- create a service which create a task in an existing project;
- create a sale order with this service;
- confirm the quotation to automatically create a related task;
- make sure the stage of the sale order is in draft or sent (cancel -> set to quotation);
- change the person assigned to the created task;
- confirm the new quotation.

Issues:
The saleperson of the sale order is also changed even if the project is not a "fsm project".

Causes:
the field service application (and more precisely the industry_fsm_stock module) wants the salesperson to be the person assigned to the task (for example, to track a stock)

Solutions:
Check the type of the project which should be a "fsm project" to change the salesperson.

opw-2963865

closes odoo/enterprise#33833

X-original-commit: 66996ab011485f213220d9fbf17f31b799ed3b24
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>