[FIX] web_studio: knowledge should not be editable with studio
Before this commit, the knowledge actions were editable with studio.
In the worst case scenario, it lead to crashes, in the best, it lead to editing
a view which made little sense.

After this commit, for simplicity's sake, we forbid that any action that presents
the knowledge.article model be editable in studio.

closes odoo/enterprise#32817

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>