[FIX] account_reports: fix multiple bugs for the accounts coverage report
1.
Currently if two expressions use the same accounts but on different date scopes, a false warning will be triggered as they appear as duplicated.

For example, this is often the case when we have a line "Previous Years Profit" in the Balance Sheet that cross references the value of the Profit & Loss report. These two duplicated expressions will match the same accounts, however they will have different date scope resulting in a false warning.

To fix this, we add a condition based on the date_scope before marking lines as candidate duplicates.

2.
Same as 1. but for subformula

3.
The excluded account codes were not correctly matched with a wildcard.
I.e. `['!', ('code', '=like', '1%')]` should be used instead of `[('code', 'not like', '1%')]`

4.
The following domain would crash
`['&', ('debit', '>', 0.0), ('account_id.code', '=like', '2679%')]`
This is because we skip tuples that do not target the `account_id`.
However if the tuple is precedeed by an unary operator, we must also skip it, otherwise we would have an incorrect domain.

5.
Fix a missing unpack star for the common domain for other reports than the BS or P&L

closes odoo/enterprise#32821

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>