[FIX] social{_sale}: disable engagement stat button.
In the new owl framework, buttons need to have a type
defined (object + method in the model code for instance).
In the social post form view, we show the engagement stat
button. However, this is only informative, it does not do
anything as a button. Therefore, when clicking it, it
triggers a traceback (no type defined etc).

In order to still show the button but to avoid creating an
empty action, the stat button is still showed but disabled.
This matches what is done in the campaign form.

We remove the name and update the related xpath in
social_sale.

Task-3016141

Part-of: odoo/enterprise#32715