[FIX] sale_subscription: assign subscription to payment.transaction in portal
before this commit, transactions were created without the sale_order_ids value.
The side effect of this is transaction references falling back on the tx-2022
format, but there may be other consequences such as subscriptions not being
linked to their transactions, payment confirmation messages not being posted
on subscriptions' chatters, making payment reconciliation more complicated, etc.

Moreover, the test_automatic_invoice_token was relying on manual patch start and
stop which is a bad practice. A context manager has the advantage to stop patching
the method if something goes wrong during the test. The patch would remain
if patcher.stop() was not called.

closes odoo/enterprise#32847

X-original-commit: 3164f15c035ff4f3f0799590a516b3c36bc45431
Signed-off-by: Arnaud Joset <arj@odoo.com>