[FIX] web_enterprise: disable home-menu focus on mobile
Previous fix in odoo/enterprise@2361672445286a4cea737aec3640e66e3894abba
introduces autofocus on home-menu, because on desktop we want to
quickly search a menu, see `o_menu_search_input` but this feature
doesn't exist on mobile. So we have to fix this because setting the
focus on this input opens the virtual keyboard.

Note that we doesn't have this issues in Odoo 13 thanks to `d-none` on
the parent element.

Steps to reproduce:
- On mobile device, open an app
- Click on the home-menu icon
- unwanted virtual keyboard appears

closes odoo/enterprise#32881

X-original-commit: a912cd5f16aebd795e712fa947c37c5657cf0f8a
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Romain Estievenart (res) <res@odoo.com>