[FIX] hr_contract_salary: Display warning when employee mail is not set
In case HR provides the simulation link to the employee via copy pasting
the link from the wizard, and employee does not have email set on home
address, after filling in the configurator, it will raise an error that
not all signers have valid email address.
To avoid this behavior, we display warning in the wizard.

task -2971112

closes odoo/enterprise#33996

X-original-commit: 520b333c8326e2e5e743edbdc7f43017b1e40890
Signed-off-by: Kevin Baptiste <kba@odoo.com>