[FIX] account_accountant: perf of rec widget when partner set
Since we already know the id, we shouldn't have to do a ilike on the
whole table of `res.partner` when not needed.
The search was previously done through the `filter_domain` on
`<account.move.line>.name`, which adds a search on all accounts and all
partners.
This is still the default behavior when searching in the list view,
which is not too annoying because we will have to search for it anyways
if we don't know the `id` beforehand -- which is not the case here.

Part-of: odoo/enterprise#32873