[FIX] account_reports: change searchbar
The filtering logic was previously fixed in [commit 1] but was broken in [commit 2].

For report containing a "searchKey" (ex: General ledger where searchKey is in fact the accountCode), the search bar only looked at the searchKey and never went through the "else" statement to search on the name. It is now corrected by searching first on the code and then on the name. For other report that don't have a "searchKey", a search on the displayName is done.

Another problem was that the total rows were displayed in the search because a total row has a displayName looking like "total 123 accountName". When it's the case, the accountName variable was "123 accountName", now the total rows are skipped to avoid processing them in the search.

This search method is now the same as v15.0, and it also restores the logic of the export and print functionality.

In addition, a tour has been added to verify that the search works when searching for a code or name.

[commit 1]: https://github.com/odoo/enterprise/commit/e7e113e5a26b357da5c01caf9a278c5dfada0366
[commit 2]: https://github.com/odoo/enterprise/commit/f435654f70f3d6a1fec7fa696945f2c0eb9a7056

closes odoo/enterprise#34538

X-original-commit: 10ce08340e12e9d57426a8e05332dcf1238f1e18
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Co-authored-by: Tom De Caluwé <decaluwe.t@gmail.com>