[IMP] account_followup: improve the automatic setting of the next followup action date
Currently, when a next followup action date is set, it depends on the delay of the next followup level. If there is no next level, the date is arbitrarily set in 14 days. This is hard coded and doesn't depend on anything (aka it's weird and bad).

This commit improves the setting of this next followup action date, by handling the different cases we can encounter:
- by default (current behavior as well) -> next date set in (next level delay - current level delay) days
- no next level -> next date set in (current level delay - previous level delay) days
- no next level AND no previous level -> next date set in (current level delay) days
- no level defined at all -> next date not set
Also updates the tooltip to better explain this process.

task id=3012793

closes odoo/enterprise#32892

Signed-off-by: Laurent Smet <las@odoo.com>