[ADD] knowledge: conversion of article to item and vice-versa
This commit enables the users to convert any children article to an
item, which can be seen inside of a kanban or list view, or any item to
an article.

This commit uses a non-compliant way to create event for buttons inside
of a view form. This will be refactored in a future task in order
to comply with the JS Framework.

task-3012915

closes odoo/enterprise#32883

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>