[FIX] mail_enterprise: run chat expand tour without outdated notifs
Before this commit, TestMailChannelExpand could fail
non-deterministically:
https://runbot.odoo.com/runbot/build/20182035

This happens because the tour could be in a state where it fetches
channel being in a folded state, but then it replays outdated
notifications that tells the channel is closed and then folded.

If unlucky, tour might click on the "expand" button of the chat
window when the chat window state transition from folded to closed,
but hasn't replayed the transition to "folded". In this precise
scenario, the click expand is ignored because the chat window does
not exist.

Note that chat window on UI is about to be removed, but it happens
every animation frame, not immediately from a change of
state. This explains why it's possible to click on "expand" button
althougth there could be no chat window. This behaviour is correct.

The problem is the "replaying" of outdated notifications. This should
be solved in the implementation of bus, but in the mean time we fix
the tour by clearing all bus notifications.

closes odoo/enterprise#32878

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>