[FIX] account_asset: fix move display name
Steps to reproduce:
- Install Accounting App.
- Go to the Accounting App.
- Click on 'Accounting' -> 'Assets'.
- Create a new asset and confirm it.
- Click on 'Modify Depreciation'.
- Select the 'Sell' action.
- Click on the empty 'Customer Invoice' field.

Observed behavior:
The invoices are listed based on their name, which is not very explicit/helpful.

Desired behavior:
In addition to the name, the partner and the date are also displayed in the list.

The search is also improved, in order to use the partner name
(in addition to the invoice name) to filter invoices.

closes odoo/enterprise#33026

X-original-commit: 039e873776496aad933062bd60febbdfcc676209
Related: odoo/odoo#103696
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>