[FIX] timer: avoid flicker on the timer when resume timer
Before this commit, when the user stops the timer and resume the timer,
the timer flicker, as if there were 2 timers running at the same time.
The reason is because the `onWillUpdateProps` is called 2 times on the
row (once since the value of timer_start has changed and another since
timer_pause changed) because of that, since `onWillUpdateProps` is a
async method, it seems the method starts 2 timers.

This commit avoids that by calculating the server offset in
`onWillStart` hook and then the `onWillUpdateProps` will no longer be a
async hook since no rpc will be called anymore.

closes odoo/odoo#103196

closes odoo/enterprise#33212

X-original-commit: d397ef5d37471fc54b22ca46d9cfe99af63360cb
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Xavier <xbo@odoo.com>