[FIX] mrp_mps: avoid replenish traceback
When the mouse pointer enter or leave, the main "REPLENSIH" button
or the row-specific "REPLENISH" link, it triggers an event that
fetch elements of the row(s) based on a specific class attribute.
However, some row(s) of the MPS don't have any element that refer this
specific class and so, trying to access the first index of the fetched
result throw a TypeError: cannot read properties of null.

closes odoo/enterprise#32893

Task: 2985735
Signed-off-by: Tiffany Chang <tic@odoo.com>