[FIX] iot: Fix download iot_handlers
In odoo there are some JS modules which may not have a path.
This is the case of the module created by studio for customizations.

Currently we are looking for the paths of all installed modules
to check if there is an iot_handler folder
But if it's a JS modules, it's impossible to find a folder and
the controller returns an error.

We now check that there is a path to the module before trying
to see if there is an iot_handler folder in it

closes odoo/enterprise#32895

Signed-off-by: Masereel Pierre <pim@odoo.com>