[IMP] account_accountant: bank recon performance (reduce search_read)
The web framework / owl automatically fetches records from the server when required. It is done in the low-level useModel hook, that is used by most views.
Since the Account Move Line List (and batch list) are full list views embedded inside a form view - any rerender to the form view will re-render child components and therefore trigger the model.load function.
On larger databases, fetching the same list of records can have a significant impact on performance.

To avoid this unneccessary call, an override of the relational model is used. A condition (the domain has changed) is applied before the reload of records, hence reducing the number of web_search_reads. This also ensures that users are still able to search for move lines and batches.

Part-of: odoo/enterprise#32898