[FIX] l10n_nl_reports: fix perf and MemoryError for XAF export
When generating the general ledger and exporting it in the XAF format,
a MemoryError could occur because of the sheer number of rows that the
get_vals_dict query produced. This was mostly because of the
LEFT JOINs on res_partner_bank and res_bank that led to
lots of duplicated rows with the same move_line_id, move_id,
journal_id, tax_id, partner_id, account_id.

Because we functionnaly only need one res_partner_bank for each partner_id,
this commit replaces the two LEFT JOINs with a single LEFT JOIN on
a CTE. There is a DISTINCT ON (partner_id, company_id) inside this CTE
so that the resulting set only contains one row for each (partner_id,
company_id) pair. That way we ensure that the LEFT JOIN will
only produce one row by account_move_line.

This does not only fix a possible MemoryError getting raised, it also
fixes a performance edge case with a small date range of
account_move_line.date. Briefly, before this patch, a small date range
on a big database could lead postgres to wrongly plan the joins order
thanks to a skewed output rows estimate. Because of that postgres
produced way too many rows after all the JOINs. This lead to the sort
algorithm being an external sort (sort on disk) instead of an
in-memory merge sort as sorting in-memory would overflow the work_mem.
This made the whole query around 20 times slower for small (1-week)
date ranges.

closes odoo/enterprise#33136

X-original-commit: efc402eb44a791c9b449d2fccf1308630c73a09e
Signed-off-by: John Laterre (jol) <jol@odoo.com>