[FIX] account_reports: fix journal report drill down tax grids
Steps to reproduce:
- Install Accounting App.
- Make sure at least one invoice has a PDF attachment.
- In Accounting, go to Reporting -> Journal Report.
- Inside the Taxes Applied box, click on a tax name.
- In the list view, click on the line of the invoice that has a PDF.

Observed behavior:
It opens the form view of the invoice.

Desired behavior:
It should open the attachment, if any, or do nothing.
Indeed, opening the form view does not provide any additional/useful
information that is not already available in the list view.
What is interesting is to have an overview of the PDF document.

closes odoo/enterprise#32902

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>