[FIX] planning: fix onboarding tour for planning app
Before this commit, when the user selects a role in his new shift, the
step of the tour keep blocking on the role selection and so the
onboarding does not go to the next step even if the user has selected or
created a role for his new shift.

This commit fixes the issue by updating the trigger used in that step to
correctly go to the next step when the `role_id` field is filled in.

Part-of: odoo/enterprise#32920