[FIX] industry_fsm_sale: remove the step to click on save button
Before this commit, the onboarding tour of Field Service has always a
step to click on the save button when the user has finished to create a
new product. This step is useless since the always edit is there and so
when the user will exit the view to go back on the kanban view of
products, the form view will be saved.

This commit removes the step clicking on the save button when the user
finished to create a new product. Also, some steps are updated to
correctly click on the breadcumb to go back to the previous view.

closes odoo/enterprise#32920

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>