[IMP] helpdesk_sale_timesheet: display SOL field for timesheets items in the portal view of a ticket
helpdesk.ticket portal form view : The SOL field is displayed in the list of timesheets entries if some of those entries do not have the same SOL as the one set on the ticket. All SOL fields are made clickable and redirecting to the corresponding portal view.

related: odoo/odoo#103434
task-3024151

closes odoo/enterprise#32860

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>