[FIX] social_push_notifications: remove o2m from group tag
The kanban view of the push notifications seems broken: the
cards are super small in terms of width.

This is because of the new grid display. The group tag should
not be around the soial post one2m and creates undesired behavior.

Removing the tag solves the issue.

Task-3016141

X-original-commit: f2d564916973a9278f8ef6d94f58aaf510ee1072
Part-of: odoo/enterprise#33504