[FIX] hr_contract_salary: set contract correctly on simulation link
Sending a simulation link currently recomputes the contract_id on the
wizard, meaning that you can't manually change which contract to use.
This is caused by the _compute_from_job method that recomputes on
every record save. This commit extracts the part that computes the
warning message from this method into _compute_warning_message, so
that this is no longer the case.

task-3001402

closes odoo/enterprise#33947

X-original-commit: d0460222f152ee351a40af6e514fd734073ddc46
Signed-off-by: Kevin Baptiste <kba@odoo.com>