[FIX] social: fix kanban view stats overflow.
In the new kanban view, the record width is set by the
kanban view. When having stats wider than the record, the
overflow is visible on the right of the card.

This is fixed by using flex wrap: if too much content is
there, it will jump to the next line. This way, we keep
readability and previous alignment. This works well
for all screen sizes.

Task-3016141

Part-of: odoo/enterprise#32715