[IMP] sale_renting: color consistency for rental status
Before this commit, the color badge for rental orders had a different
color, depending on the view, leading to inconsistencies.

After this commit, rental badges are uniformized between views and with
the line color in the tree view.

The default string for a confirmed rental order has been changed to
"booked" to reflect the rental status.

task-3010707

closes odoo/enterprise#34188

Related: odoo/odoo#106843
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>