[FIX] l10n_cl_edi_stock: Equalized delivery guide XML dates so that they always match
The following fields in the XML now all use the current timestamp
<TSTED>, <TmstFirma>, <TmstFirmaEnv>, <FchEmis>, <FE>

Original Issue:
Delivery guides are being rejected because it has 2 dates. One that comes from the expected
delivery from the Sales Order and the one of the date of the stamp. If these 2 dates differ, the
SII rejects the delivery guide. We agreed that making every date the current timestamp would
make the most sense businesswise because making it dependent on other fields could indirectly
make users set the wrong dates on the stock.picking, which could impact KPIs.

closes odoo/enterprise#34617

X-original-commit: 3b0782fceb9b9b4baf8d0bb9884c8c06786e6393
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>