[FIX] social,appointment: restore "onchange_on_keydown" feature
Previously, we had an option on char/text fields called "onchange_on_keydown",
which allowed to trigger the onchange while the user was typing (instead of
when the field is blurred).

It was notably used in the following contexts:

- When typing a message in the social app, to get the previews
- When typing a Youtube title/description in the social app, to update the
  Youtube preview
- When modifying the appointment share link "short code", to warn the user that
  he cannot enter special characters (such as a space)

This feature was lost during the port to OWL, this commit restores it.

Task-3033107

closes odoo/enterprise#34211

X-original-commit: ffc625333940301cb12d1686ea7bddd8617095da
Related: odoo/odoo#106193
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>