[FIX] hr_recruitment_extract: show OCR alert when file uploading
Before this commit, when uploading a file in chatter on `hr.applicant`,
we had to manually reload the view to see alert to let us know an OCR
process is running from this file upload.

This happens because showing of alert requires reloading the view,
but file upload does not reload view. This is intended behaviour for
most view, as it's unnecessary.

In the case of `hr.applicant`, we want to reload view.

This commit fixes the problem by patching triggering rule to reload
view of chatter from a file upload, so that it always does it when
it's a view on `hr.applicant`.

Note that this solution is specific to `hr.applicant`. So other
models with OCR process in background won't work until they add a
similar patch.

As future work, would be nice to have awareness of presence of OCR
process in background, so that code relies instead on this
information and it will work for any views that has this future.

Task-2886634

closes odoo/enterprise#32982

Related: odoo/odoo#103557
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>