[FIX] account_reports: fix _init_custom_options()
Currently, when opening a variant (aka not a root report),
the options are not set properly.

That is because _init_custom_options() will try to call
_custom_options_initializer() from the custom handler.

But if the variant does not have a custom handler,
or if the method is not defined on the variant's handler,
we should try to call _custom_options_initializer() on
the root report's custom handler, which was not the case.

This PR fixes the issue.

closes odoo/enterprise#32984

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>