[FIX] planning: take availabilities into account in Gantt totals
This commit ensures that the resource availabilities are taken into account
while computing gantt totals. This means that if we consider being in the
week scale, a pill that would cross the whole week will not be taken into
account in the totals for the period where the associated resource is not
supposed to work.

This commit also ensures that total pills height is related to the number
of allocated hours instead of the number of pills.

task-2971118

closes odoo/enterprise#31032

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: Laurent Stukkens (LTU) <ltu@odoo.com>