[FIX] (l10n_be_)hr_contract_salary, l10n_be_hr_payroll: adjust cost_res_field_public and res_field_public selections
The - cost_res_field_public and res_field_public - are selection fields
consisting of 'hr.contract' fields. By default non-stored fields are
not in the selection, unless we add them via _advantage_white_list method.

Prior, some of the computed non stored fields that are used in the Advantages were
not added explicitly and it caused traceback when we clicked on the
records of those advantages that have non-stored fields set for
cost_res_field_id and res_field_id.

Moreover, prior wage_with_holidays was backlisted. it causes issues for
the advantage "13th Month", where res_field_id is set to wage_with_holidays
and as the field is blacklisted we have issues when res_field_public is
computed as the field is not in the selection. The field wage_with_holidays
was blacklisted prior to avoid recomputation of the _compute_final_yearly_costs.

Now it is possible to set 'wage_with_holidays' as res_field_id without
causing extra computation.

closes odoo/enterprise#33002

Signed-off-by: Gvaladze Sopiko (sgv) <sgv@odoo.com>