[FIX] mrp_workorder: display worksheet when no step
Originally, if there's a worksheet but no step set on an operation, the worksheet is not displayed in the tablet view.
Even if the 'get worksheetData()' method in tablet.js is made to handle the case where there's no step,
it was never called from the xml as it was required to have a step there.
The DocumentViewer block is moved in parent to solve this issue, however, as we don't want to show the DocumentViewer
in same time as the SummaryStep, the display condition in the xml evolve a little.

closes odoo/enterprise#33001

Task: 2985735
Signed-off-by: Tiffany Chang <tic@odoo.com>