[FIX] data_merge: speed up _search_company_id on big databases.
Because of the multi-company record rule on data_merge_record, each
search on env['data_merge.record'] (without sudo) will call
_search_company_id. That's because data_merge.record.company_id
is a non-stored computed field. This method fetches all
records (active_test=False) on the data_merge_record table
then calls filtered on the resulting recordset. This induces
two performance bottlenecks. One, a filtered on a potentially
huge amount of records (data_merge_record can grow quite big)
can lead to slowdowns. Two, since company_id is a computed
field, accessing this field inside the filtered lambda function
means to recompute the field value. So, if we are filtering
300K records on company_id, the corresponding compute function
will be called 300 times.

Optimizing that in stable is not really straightforward,
as the easiest solution (done in master odoo/enterprise#33768)
would be to store the company_id field. Here we avoid calling the compute
function of company_id and instead we "manually" build a subquery
by combining custom SQL and Query objects constructed with the
_search method. Injecting the resulting Query in a domain
of the form [('id', 'in', Query)] leads to significant speedups
in large databases.

opw-2968536

closes odoo/enterprise#35165

X-original-commit: 3cb54409543816affb173396f756a7b3ec158480
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Van Delft Aurélien (avd) <avd@odoo.com>