[FIX] (website_)appointment: fix picture flicker
In appointment, images flicker because of the use of an HTML <img> tag. This is
because the image is first loaded and displayed in the IMG tag, then styles are
applied and the image is redimensionned.

To avoid this, a solution is to use a div with a background image - because
everything happens in CSS, the image will not be displayed until its style has
first been calculated. The use of a background image is valid here because the
pictures are not editable from those pages.

Task-3002352

closes odoo/enterprise#33111

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>